Rename "files" module to "extractors"
